backend/AI/PictureDetect/requirements.txt
flask
pillow
numpy
keras==3.11.3
tensorflow==2.20.0